+++ libs/base_core/src/sys/mem/byte.rs
@@ -5,11 +5,13 @@
 // Used by:
 // - sys::mem::arena::ArenaBytes.
 
+use crate::MaybeUninit;
+
 crate::CONST! {
     _DOC_MAYBE_BYTE = "A byte type that may be uninitialized depending on features.
 
 # Features
-When `unsafe_array` is **enabled** it aliases [`crate::MaybeUninit<u8>`].
+When `unsafe_array` is **enabled** it aliases [`MaybeUninit<u8>`].
 Otherwise it aliases [`u8`].";
 }
 
@@ -19,4 +21,4 @@ pub type MaybeByte = u8;
 
 #[doc = _DOC_MAYBE_BYTE!()]
 #[cfg(all(not(base_safe_mem), feature = "unsafe_array"))] // unsafe
-pub type MaybeByte = crate::MaybeUninit<u8>;
+pub type MaybeByte = MaybeUninit<u8>;